Fixes #3: --only-collection exports only first collection I have fixed this bug, but I do however think the whole current approach of the script is off. Currently it is trying to make a new collection for every single resource group, which seems a bit OTT. I work on multiple APIs and each one of those is its own collection, and not just each resource for a specific API.
With this PR, the  will merge all collections into one, and the name will just be whatever the name of the first resource is. I'd prefer to use the name form the first environment, as you've done so far with other parts, and I would like to see the whole script treat each import as one collection.

Let me know if I can help with any of that, or if I can improve this PR.